@@ -8,10 +8,11 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "autoprefixer": "^10.4.2",
+    "framer-motion": "^11.15.0",
     "gh-pages": "^5.0.0",
     "postcss": "^8.4.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-icons": "^4.7.1",
     "react-scripts": "5.0.0",
     "react-scroll": "^1.8.6",
