@@ -12,7 +12,7 @@ import { HiOutlineMail } from 'react-icons/hi';
 import { BsFillPersonLinesFill } from 'react-icons/bs';
 import Logo from '../assets/logo1.png';
 import { Link } from 'react-scroll';
-import CV from '../assets/NimeshCVV.png';
+import CV from '../assets/Nimesh_CV.pdf';
 
 const Navbar = () => {
   const [nav, setNav] = useState(false);
@@ -21,7 +21,7 @@ const Navbar = () => {
   return (
     <div className='fixed w-full h-[80px] flex justify-between items-center px-4 bg-[#0a192f] text-gray-300'>
       <div>
-        <img src={Logo} alt='Logo Image' style={{ width: '175px' }} />
+        {/* <img src={Logo} alt='Logo Image' style={{ width: '175px' }} /> */}
       </div>
 
       {/* menu */}
@@ -152,7 +152,7 @@ const Navbar = () => {
           </li>
 
           <li className='w-[160px] h-[60px] flex justify-between items-center ml-[-100px] hover:ml-[-10px] duration-300 bg-[#565f69]'>
-          <a className='flex justify-between items-center w-full text-gray-300' href={CV} target="" rel="noopener noreferrer">
+          <a className='flex justify-between items-center w-full text-gray-300' href={CV} target="_blank" rel="noopener noreferrer">
   Resume <BsFillPersonLinesFill size={30} />
 </a>
 
